requirements.txt
pycryptodome==3.12.0
bcrypt==3.2.0
Pillow==8.4.0